backend/requirements.txt
python-decouple
Django
Psycopg2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
faker
django-cors-headers
python-dotenv
pymongo
django-extensions